split bgpmap into crate
