+++ route_verification/bgp/Cargo.toml
@@ -13,6 +13,7 @@ rayon = "1.7"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 
+map = { package = "route_verification_bgpmap", path = "../bgpmap" }
 bloom = { package = "route_verification_bloom", path = "../bloom" }
 io = { package = "route_verification_io", path = "../io" }
 lex = { package = "route_verification_lex", path = "../lex" }
