+++ route_verification/bgpmap/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "route_verification_bgpmap"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0"
+ipnet = { version = "2.8", features = ["serde"] }
+lazy-regex = "3.0"
+serde = { version = "1.0", features = ["derive"] }
